requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langchainhub
sentence-transformers
faiss-cpu
streamlit
pypdf
pyrebase4
beautifulsoup4
